Resync until last event
This pull request improves client resync logic so resync is called until the last event is returned or ResyncEventsThreshold is reached. Previously if a client had a strict filter applied and ResyncFn implementation had a limit on returned events it was possible for clients to fail resyncing to get any possible new events. ResyncFn has also been changed to return an error when resync failed and connection should be closed. ResyncFn no longer needs to return an ok argument if not all events could be retrieved. It is now this library’s responsibility to make sure resync completes fully.

Approved-by: Julius Kriukas <julius.kriukas@adv.bet>