AADV-1000: Cached count stream
This pull request introduces a new CachedCountStream type which can be used for cached streams with a set limit of cached events instead of current CachedStream implementation, where expiration time is used.

When cached events count passed the set size limit the oldest event is removed to make place for a new event.

ErrCacheMiss error is returned when client cannot resync due to missing events in the cache.

Approved-by: Dainius Grėbliauskas <dainius.grebliauskas@adv.bet>