Always send past events if the given lastEventID is an empty string
If the internal lastEventID was never updated (left as a default string value) passing an empty string as the `lastEventID` would not publish last events because the of the `If` statement.
This change adds to the `If` statement that if the `lastEventID == ""` we should always send past events if we have any.

Approved-by: Julius Kriukas <julius.kriukas@adv.bet>