Fix: Use resync limit in cache implementation
Not using a re-sync limit can cause problems when the clients become delayed, in a one connection it tries to re-stream all missing cache events, which requires a high memory allocation and a way higher loading times (thousands of the events are being retrieved from the cache, which contains a mutex). The same re-sync events threshold logic is used in the generic stream. If the client has an unstable connection, with a higher delay and missing events count, the re-sync procedure can become harder and more costlier.

This cache implementation is used in live and prematch server sent events streaming logic.